move port acquisition to package
