+++ cmd/dev/main.go
@@ -11,6 +11,7 @@ import (
 
 	version "github.com/caarlos0/go-version"
 	"github.com/hugginsio/dev/handler"
+	"github.com/hugginsio/dev/port"
 )
 
 type Config struct {
@@ -27,7 +28,7 @@ func main() {
 		os.Exit(http.StatusBadRequest)
 	}
 
-	port, err := acquirePort(config.port)
+	port, err := port.Acquire(config.port)
 	if err != nil {
 		slog.Error("failed to find available port", "err", err.Error())
 		os.Exit(http.StatusInternalServerError)
